llm_retrieval/requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
huggingface-hub>=0.16.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
nltk>=3.8.1
numpy>=1.23.0
sacrebleu>=2.0.0
colorama>=0.4.4
unbabel-comet
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
plotly>=5.0.0